1.0.3
Fixed some typos in Autohotkey shortcuts and vassal contract window.
Added Shift F one and Shift F2 as replacements for Page Up and Page down, if you don't have them.
Mentioned the new hotkeys in the tutorial.
In the title window, made the whole line with the de jure parent clickable.
Restored the default fonts for German, French and Spanish to get the accents back.