1.0.4
Added 2 more rings of adjacent counties and a directional filter for them.
Added holders of adjacent counties.
Added a toggle to not auto-close adjacent rulers.
Added a sound when someone joins your multiplayer game.
Added debug info in the character interactions menu.
For dead characters, added the debug tooltip to the word dead.