1.1.0
Updated for 1.12 Scythe patch and Legends of the Dead DLC.
Added legitimacy to resources, after piety. In the character window it's hotkeyed to L.
Added plague and legends buttons, hotkeyed to P and L on the main screen.
Fixed: Tax jurisdictions button on the main screen is now enabled for all rulers who have clan vassals.
Fixed renaming artifacts.